+++ AYIM/api/index.ts
@@ -31,7 +31,7 @@ app.koa.use(mount("/staff/users", usersRouter.routes()));
 
 app.koa.use(mount("/admin", adminRouter.routes()));
 app.koa.use(mount("/admin/results", adminResultsRouter.routes()));
-app.koa.use(mount("/admin/categories", adminCategoriesRouter.routes()));
+app.koa.use(mount("/admin/years", adminCategoriesRouter.routes()));
 app.koa.use(mount("/admin/years", adminYearsRouter.routes()));
 
 export default {
