--- MCA/pages/_year/admin/index.vue
@@ -1,29 +1,0 @@
-<template>
-    <div>
-        <mode-switcher :hide-phase="true">
-            <admin-page />
-        </mode-switcher>
-    </div>
-</template>
-
-<script lang="ts">
-import { Vue, Component } from "vue-property-decorator";
-
-import AdminPage from "../../../../MCA-AYIM/components/admin/AdminPage.vue";
-import ModeSwitcher from "../../../../MCA-AYIM/components/ModeSwitcher.vue";
-
-@Component({
-    components: {
-        ModeSwitcher,
-        AdminPage,
-    },
-    head () {
-        return {
-            title: "Admin | MCA",
-        };
-    },
-})
-export default class Admin extends Vue {
-    
-}
-</script>
